pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mergerfs-balance"
version = "0.1.0"
description = "Parallel drive balancer for mergerfs filesystems"
requires-python = ">=3.9"
license = "MIT"
authors = [
    { name = "Adam Newell" }
]
keywords = ["mergerfs", "balance", "filesystem", "storage"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Filesystems",
]
dependencies = [
    "rich>=13.0.0",
]

[project.optional-dependencies]
yaml = ["pyyaml>=6.0"]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

[project.scripts]
mergerfs-balance = "mergerfs_balance.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/mergerfs_balance"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
target-version = "py39"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]
ignore = ["E501"]  # line length handled by formatter

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.9"
strict = true
warn_return_any = true
warn_unused_ignores = true
ignore_missing_imports = true
